pyproject.toml
[tool.poetry]
name = "radeky_bot"
version = "0.4.0"
description = "radeky_bot"
authors = ["Radekyspec <w41k3r15347@gmail.com>"]
readme = "README.md"
packages = [
    { include = "radeky_bot", from = "./src/plugins/"}
]

[tool.poetry.dependencies]
aiofiles = "^0.8.0"
aiohttp = "^3.8.1"
brotlipy = "^0.7.0"
nb-cli = "^0.6.5"
nonebot-adapter-onebot = "^2.0.0b1"
nonebot-plugin-apscheduler = "^0.1.2"
pyyaml = "^5.4.1"
rsa = "^4.8"
DateTime = "^4.3"
websockets = "^10.1"
pydantic = "^1.9.0"
pycryptodome = "^3.14.0"


[tool.poetry.dev-dependencies]
aiofiles = "^0.8.0"
aiohttp = "^3.8.1"
brotlipy = "^0.7.0"
nb-cli = "^0.6.5"
nonebot-adapter-onebot = "^2.0.0b1"
nonebot-plugin-apscheduler = "^0.1.2"
pyyaml = "^5.4.1"
rsa = "^4.8"
DateTime = "^4.3"
websockets = "^10.1"
pydantic = "^1.9.0"
pycryptodome = "^3.14.0"


[nonebot.plugins]
plugins = []
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
